backend/pyproject.toml
[tool.poetry]
name = "blocknotify-net"
version = "0.1.0"
description = "A notification service that monitors blockchain events."
authors = ["Nik <nik@pyx.is>"]
readme = "README.md"
#packages = [{include = "blocknotify"}]
license = "MIT"


[tool.poetry.dependencies]
python = "^3.8"
django = "^4.1.7"
dj-database-url = "^1.2.0"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.14.0"
#ninja = "^1.11.1"
django-ninja = "^0.21.0"
eth-keys = "^0.4.0"
eth-hash = "^0.5.1"
pycryptodome = "^3.17"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
